node/mempool_precheck: PR #1422 wave-10 split into 3 files (Codacy file-budget — input + output + orchestration)
Wave-10 closes the wave-9 Codacy file-cumulative complexity warning
by splitting mempool_precheck.go into three sibling files in the same
package, each below the per-file complexity budget:

- mempool_precheck.go (orchestration): buildPolicyInputSnapshotIfNeeded,
  checkTxParseAndContext, checkTransactionWithSnapshot,
  cheapFeeFloorPrecheck, precheckEarlyDefer
- mempool_precheck_input.go (input-side helpers):
  feePrecheckP2PKInputValue, precheckP2PKInputStructurallyValid,
  precheckCoinbaseImmature
- mempool_precheck_output.go (output-side helper):
  feePrecheckP2PKOutputValue

Same package, no API change, no test impact. Each file's cumulative
complexity now within the repository's per-file budget.

Wave-10 cumulative gate verification:
- gofmt -l ./node clean
- go vet ./node/... clean
- go build ./node/... clean
- go test ./node -run TestMempool -count=1 PASS
