+++ clients/go/node/mempool_precheck.go
@@ -2,7 +2,6 @@ package node
 
 import (
 	"fmt"
-	"math/bits"
 
 	"github.com/2tbmz9y2xt-lang/rubin-protocol/clients/go/consensus"
 )
@@ -173,126 +172,4 @@ func precheckEarlyDefer(tx *consensus.Tx) bool {
 		return true
 	}
 	return tx.TxNonce == 0
-}
-
-// feePrecheckP2PKInputValue returns the P2PK input value when tx has
-// exactly one input AND that input is BOTH structurally valid
-// (witness count == 1, no coinbase-prevout marker, empty ScriptSig,
-// Sequence in standard range) AND resolves in utxos to a
-// COV_TYPE_P2PK entry. Returns (0, false) for any other shape so the
-// caller defers to the expensive admission path.
-//
-// Wave-4 class-closure conservatism: each input-side guard mirrors a
-// terminal-reject branch in the slow path
-// `applyNonCoinbaseTxBasic*` at
-// `clients/go/consensus/utxo_basic.go:193-200`. Without these defers
-// a below-floor tx with structurally-defective input would be
-// misclassified as transient Unavailable instead of terminal
-// Rejected, masking the structural error and allowing callers to
-// retry forever. Mirrors Rust `fee_precheck_p2pk_input_value` in
-// `clients/rust/crates/rubin-node/src/txpool.rs`.
-//
-// Scope-cap: P2PK signature-verification failure is NOT classified
-// here. Verifying ML-DSA signatures is the expensive operation this
-// fast-reject is designed to avoid. Below-floor txs with invalid
-// signatures may surface as rolling-floor Unavailable until the fee
-// floor no longer applies; this is intentional.
-func feePrecheckP2PKInputValue(tx *consensus.Tx, utxos map[consensus.Outpoint]consensus.UtxoEntry, nextHeight uint64) (uint64, bool) {
-	if utxos == nil || len(tx.Inputs) != 1 || len(tx.Witness) != 1 {
-		return 0, false
-	}
-	in := tx.Inputs[0]
-	if !precheckP2PKInputStructurallyValid(in) {
-		return 0, false
-	}
-	entry, ok := utxos[consensus.Outpoint{Txid: in.PrevTxid, Vout: in.PrevVout}]
-	if !ok || entry.CovenantType != consensus.COV_TYPE_P2PK {
-		return 0, false
-	}
-	if precheckCoinbaseImmature(entry, nextHeight) {
-		return 0, false
-	}
-	return entry.Value, true
-}
-
-// precheckP2PKInputStructurallyValid returns true iff the input is
-// structurally valid for the cheap P2PK precheck. Wave-4 guards mirror
-// terminal-reject parse-time checks at clients/go/consensus/utxo_basic.go
-// for non-empty ScriptSig (193-194), out-of-range Sequence (196-197),
-// and coinbase-prevout marker on non-coinbase (199-200). Returns false
-// on any structural defect so the caller defers to the slow path.
-func precheckP2PKInputStructurallyValid(in consensus.TxInput) bool {
-	var zeroTxid [32]byte
-	if in.PrevTxid == zeroTxid && in.PrevVout == 0xffffffff {
-		return false
-	}
-	if len(in.ScriptSig) != 0 {
-		return false
-	}
-	if in.Sequence > 0x7fffffff {
-		return false
-	}
-	return true
-}
-
-// precheckCoinbaseImmature returns true iff the resolved P2PK input is
-// an immature coinbase spend (wave-5 class-closure: slow path returns
-// Rejected TX_ERR_COINBASE_IMMATURE at utxo_basic.go:217-219). Caller
-// defers when this returns true so the slow path preserves the
-// terminal-reject classification (different caller action than fee
-// floor: wait for COINBASE_MATURITY blocks vs retry-with-higher-fee).
-func precheckCoinbaseImmature(entry consensus.UtxoEntry, nextHeight uint64) bool {
-	if !entry.CreatedByCoinbase {
-		return false
-	}
-	if nextHeight < entry.CreationHeight {
-		return true
-	}
-	return nextHeight-entry.CreationHeight < consensus.COINBASE_MATURITY
-}
-
-// feePrecheckP2PKOutputValue returns the sum of P2PK output values when
-// every output is a CONSENSUS-VALID `COV_TYPE_P2PK` (non-zero value,
-// exactly `MAX_P2PK_COVENANT_DATA == 33` byte covenant_data, and a
-// suite_id in the active `NativeCreateSuites(nextHeight)` set) and the
-// running sum does not overflow `uint64`. Returns `(0, false)` for any
-// other shape so the caller defers to the expensive admission path.
-//
-// Wave-4 class-closure conservatism: each guard mirrors a permanent-
-// reject branch in `ValidateTxCovenantsGenesis` at
-// `clients/go/consensus/covenant_genesis.go:19-26`. Without them a
-// below-floor tx with consensus-invalid P2PK outputs would be
-// misclassified as transient Unavailable instead of permanent Rejected.
-// Mirrors Rust `fee_precheck_p2pk_output_value` at
-// `clients/rust/crates/rubin-node/src/txpool.rs`.
-func feePrecheckP2PKOutputValue(outputs []consensus.TxOutput, nextHeight uint64, rotation consensus.RotationProvider) (uint64, bool) {
-	if rotation == nil {
-		rotation = consensus.DefaultRotationProvider{}
-	}
-	nativeSuites := rotation.NativeCreateSuites(nextHeight)
-	var total uint64
-	for _, out := range outputs {
-		if out.CovenantType != consensus.COV_TYPE_P2PK {
-			return 0, false
-		}
-		// Wave-4 class-closure conservatism: each guard mirrors a
-		// permanent-reject branch in ValidateTxCovenantsGenesis
-		// (covenant_genesis.go:19-26).
-		if out.Value == 0 {
-			return 0, false
-		}
-		if uint64(len(out.CovenantData)) != consensus.MAX_P2PK_COVENANT_DATA {
-			return 0, false
-		}
-		suiteID := out.CovenantData[0]
-		if !nativeSuites.Contains(suiteID) {
-			return 0, false
-		}
-		next, carry := bits.Add64(total, out.Value, 0)
-		if carry != 0 {
-			return 0, false
-		}
-		total = next
-	}
-	return total, true
 }
